VirtIO-FS: move scope_exit to separate file
